Reusable estimator
