+++ tools/src/tensil/tools/compiler/LIREstimator.scala
@@ -3,118 +3,20 @@ package tensil.tools.compiler
 import tensil.InstructionLayout
 
 class LIREstimator(layout: InstructionLayout, stats: Stats) extends LIR {
-  private var previousOpcode = Opcode.Wait
-  private var previousFlags  = 0
-
-  private val InternalTransferEnergy = 10
-  private val InternalTransferCycles = 1
-
-  private val DRAMTransferEnergy = 100
-
-  private val DRAMTransferSetupReadCycles  = 32
-  private val DRAMTransferSetupWriteCycles = 0
-  private val DRAMTransferReadCycles       = 2
-  private val DRAMTransferWriteCycles      = 1
-  private val DRAMTransferWidthBits        = 128
-
-  private def beforeEmit(
-      currentOp: Int,
-      flags: Int = 0
-  ): Unit = {
-    previousOpcode = currentOp
-    previousFlags = flags
-  }
-
-  private def estimateCyclesAndEnergy(
-      currentOp: Int,
-      size: Option[MemoryAddressRaw] = None,
-      flags: Int = 0
-  ): (Long, Long) = {
-    currentOp match {
-      case Opcode.Wait =>
-        val cycles = 1
-        val energy = 0
-
-        (cycles, energy)
-      case Opcode.MatMul => {
-        val cycles =
-          (if (previousOpcode == Opcode.MatMul)
-             (size.get + 1)
-           else if (previousOpcode == Opcode.LoadWeights)
-             (size.get + 1 + layout.arch.arraySize)
-           else
-             (size.get + 1 + 2 * layout.arch.arraySize))
-
-        val energy =
-          (size.get + 1) * layout.arch.arraySize * layout.arch.arraySize
-
-        (cycles, energy)
-      }
-
-      case Opcode.SIMD => {
-        val cycles = 1
-        val energy = layout.arch.arraySize
-
-        (cycles, energy)
-      }
-
-      case Opcode.LoadWeights => {
-        val cycles = (size.get + 1) * InternalTransferCycles
-        val energy = (size.get + 1) * InternalTransferEnergy
-
-        (cycles, energy)
-      }
-
-      case Opcode.DataMove =>
-        if (
-          flags == DataMoveFlags.LocalToDRAM0 ||
-          flags == DataMoveFlags.LocalToDRAM1 ||
-          flags == DataMoveFlags.DRAM0ToLocal ||
-          flags == DataMoveFlags.DRAM1ToLocal
-        ) {
-          /*val transfersPerVector = divCeil(
-            dataType.sizeBytes * layout.arch.arraySize * 8,
-            DRAMTransferWidthBits
-          )
-          val isRead =
-            flags == DataMoveFlags.DRAM0ToLocal || flags == DataMoveFlags.DRAM1ToLocal
-          val transferCycles =
-            if (isRead) DRAMTransferReadCycles
-            else DRAMTransferWriteCycles
-          val setupCycles =
-            if (previousOpcode == Opcode.DataMove && previousFlags == flags) 0
-            else if (isRead) DRAMTransferSetupReadCycles
-            else DRAMTransferSetupWriteCycles
-
-          val cycles =
-            (size.get + 1) * transfersPerVector * transferCycles + setupCycles
-          val energy = (size.get + 1) * transfersPerVector * DRAMTransferEnergy*/
-
-          val cycles = (size.get + 1) * 1
-          val energy = (size.get + 1) * 1
-
-          (cycles, energy)
-        } else {
-          val cycles = (size.get + 1) * InternalTransferCycles
-          val energy = (size.get + 1) * InternalTransferEnergy
-
-          (cycles, energy)
-        }
-
-      case _ =>
-        (0, 0)
-    }
-  }
+  val estimator = new Estimator(layout)
 
   def emitNoOp(): Unit = {
-    beforeEmit(Opcode.Wait)
+    stats.countInstruction(
+      "Wait",
+      estimator.estimateCyclesAndEnergy(Opcode.Wait)
+    )
   }
 
   def emitWait(tidToWait: Int): Unit = {
-    beforeEmit(Opcode.Wait)
-
-    val (cycles, energy) = estimateCyclesAndEnergy(Opcode.Wait)
-    stats.countInstruction("Wait", cycles, energy)
+    stats.countInstruction(
+      "Wait",
+      estimator.estimateCyclesAndEnergy(Opcode.Wait)
+    )
   }
 
   def emitMatMul(
@@ -125,12 +27,13 @@ class LIREstimator(layout: InstructionLayout, stats: Stats) extends LIR {
       accumulatorAddress: MemoryAddress,
       size: MemoryAddressRaw
   ): Unit = {
-    beforeEmit(Opcode.MatMul)
-
     val mnemonic = "MatMul"
 
-    val (cycles, energy) = estimateCyclesAndEnergy(Opcode.MatMul, Some(size))
-    stats.countInstruction(mnemonic, cycles, energy, Some(size))
+    stats.countInstruction(
+      mnemonic,
+      estimator.estimateCyclesAndEnergy(Opcode.MatMul, Some(size)),
+      Some(size)
+    )
     if (localAddress.tag != MemoryTag.Zeroes)
       stats.countStride(mnemonic, 0, localStride, size)
     stats.countStride(mnemonic, 1, accumulatorStride, size)
@@ -145,10 +48,10 @@ class LIREstimator(layout: InstructionLayout, stats: Stats) extends LIR {
       writeAccumulatorAddress: MemoryAddress,
       readAccumulatorAddress: MemoryAddress
   ): Unit = {
-    beforeEmit(Opcode.SIMD)
-
-    val (cycles, energy) = estimateCyclesAndEnergy(Opcode.SIMD)
-    stats.countInstruction("SIMD", cycles, energy)
+    stats.countInstruction(
+      "SIMD",
+      estimator.estimateCyclesAndEnergy(Opcode.SIMD)
+    )
   }
 
   def emitDataMove(
@@ -161,12 +64,6 @@ class LIREstimator(layout: InstructionLayout, stats: Stats) extends LIR {
       size: MemoryAddressRaw
   ): Unit = {
     val flags = LIRGen.mkDataMoveFlags(toLocal, accumulate, address.tag)
-
-    beforeEmit(Opcode.DataMove, flags)
-
-    val (cycles, energy) =
-      estimateCyclesAndEnergy(Opcode.DataMove, Some(size), flags)
-
     val suffix = flags match {
       case DataMoveFlags.LocalToDRAM0       => "(LocalToDRAM0)"
       case DataMoveFlags.LocalToDRAM1       => "(LocalToDRAM1)"
@@ -180,7 +77,11 @@ class LIREstimator(layout: InstructionLayout, stats: Stats) extends LIR {
 
     val mnemonicWithSuffix = "DataMove" + suffix
 
-    stats.countInstruction(mnemonicWithSuffix, cycles, energy, Some(size))
+    stats.countInstruction(
+      mnemonicWithSuffix,
+      estimator.estimateCyclesAndEnergy(Opcode.DataMove, Some(size), flags),
+      Some(size)
+    )
     stats.countStride(mnemonicWithSuffix, 0, localStride, size)
     stats.countStride(mnemonicWithSuffix, 1, stride, size)
   }
@@ -190,14 +91,13 @@ class LIREstimator(layout: InstructionLayout, stats: Stats) extends LIR {
       localAddress: MemoryAddress,
       size: MemoryAddressRaw
   ): Unit = {
-    beforeEmit(Opcode.LoadWeights)
-
-    val (cycles, energy) =
-      estimateCyclesAndEnergy(Opcode.LoadWeights, Some(size))
-
     val mnemonic = "LoadWeights"
 
-    stats.countInstruction(mnemonic, cycles, energy, Some(size))
+    stats.countInstruction(
+      mnemonic,
+      estimator.estimateCyclesAndEnergy(Opcode.LoadWeights, Some(size)),
+      Some(size)
+    )
 
     if (localAddress.tag != MemoryTag.Zeroes)
       stats.countStride(mnemonic, 0, localStride, size)
